#524 Create a mixin for persisted sort
